requirements.txt
streamlit
pandas
openpyxl
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-mistralai

faiss-cpu

mistralai
Pillow
python-dotenv

requests

scikit-learn
joblib
numpy
